refactor(data): split wasm module into multiple modules
